Add solution for 2017/day20/part2
So I wanted to make an "elegant" solution where I checked the velocity
vectors of particles to determine if they may possibly collide in the future.
However, my brain decidedly refused to even try to grasp any kind of math
today, so in the end I got tired and decided to run a brute-force program.
Whaddya know, it worked. I don't feel like continuing to try to implement
the better approach, so I'm leaving the code as-is.